requirements.txt
# -----------------------
# App
# -----------------------
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# -----------------------
# PDF ingestion (Unstructured)
# NOTE: Poppler + Tesseract are SYSTEM deps (conda-forge), not pip.
# -----------------------
unstructured>=0.12.0
unstructured-inference>=0.7.0
pdf2image>=1.16.0
pillow>=9.0.0
pytesseract>=0.3.10
pi-heif>=0.13.0

# -----------------------
# Vector DB
# -----------------------
chromadb>=0.4.0

# -----------------------
# Embeddings / NLP
# -----------------------
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# -----------------------
# LLM provider
# -----------------------
groq>=0.4.0

# -----------------------
# Data utilities
# -----------------------
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# -----------------------
# Reliability
# -----------------------
tenacity>=8.2.0
